requirements.txt
# Django and Core Dependencies
Django==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
whitenoise==6.6.0
Pillow==10.2.0

# Security - Rate Limiting
django-ratelimit==4.1.0

# REST API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
